place_finder_mcp/requirements.txt

fastapi==0.115.4
uvicorn[standard]==0.32.0
starlette==0.41.3

googlemaps==4.10.0

python-dotenv==1.0.1

pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.11.0

mcp==1.18.0
fastmcp==0.3.0

requests==2.32.3
httpx==0.28.1
httpx-sse==0.4.3

loguru==0.7.2

anyio==4.11.0
jsonschema==4.25.1
